requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
langchain-huggingface
langchain-chroma
python-dotenv
langchain-google-genai
pypdf
